# api-spec/v2/openapi.yaml
openapi: 3.0.0
info:
  version: 2.0.0
  title: Pagopa eCommerce services for assistance api
  description: This microservice that expose eCommerce services for assistance api.
servers:
  - url: https://${host}
tags:
  - name: eCommerce
    description: Api's for performing transaction search on ecommerce DB
    externalDocs:
      url: https://pagopa.atlassian.net/wiki/spaces/I/pages/492339720/pagoPA+eCommerce+Design+Review
      description: Technical specifications
security:
  - ApiKeyAuth: [ ]

paths:
  /v2/ecommerce/searchTransaction:
    post:
      parameters:
        - in: query
          name: pageNumber
          schema:
            type: integer
            default: 0
          required: true
          description: "Searched page number, starting from 0"
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 10
          required: true
          description: "Max element per page"
      tags:
        - ecommerce
      operationId: ecommerceSearchTransaction
      summary: Search transaction by input parameters
      description: 'GET with body payload - no resources created'
      requestBody:
        $ref: "#/components/requestBodies/EcommerceSearchTransactionRequest"
      responses:
        '200':
          description: Transactions found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchTransactionResponse'
        '400':
          description: Formally invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
  /v2/ecommerce/searchMetrics:
    post:
      tags:
        - ecommerce
      operationId: ecommerceSearchMetrics
      summary: Search Metrics by input filters and time range
      description: >
        Search for aggregated transaction metrics based on optional filters 
        (clientId, pspId, paymentTypeCode) and a mandatory time range (max 1 hour).
      requestBody:
        $ref: "#/components/requestBodies/SearchMetricsRequest"
      responses:
        '200':
          description: Metrics found grouped by transaction status (as fixed JSON map)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionMetricsResponse'
        '400':
          description: Invalid input (e.g. missing required fields, time range > 1 hour)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '404':
          description: Metrics not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
  /v2/helpdesk/searchTransaction:
    post:
      parameters:
        - in: query
          name: pageNumber
          schema:
            type: integer
            default: 0
          required: true
          description: "Searched page number, starting from 0"
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 10
          required: true
          description: "Max element per page"
      tags:
        - helpdesk
      operationId: helpDeskSearchTransaction
      summary: Search transaction by input parameters
      description: 'GET with body payload - no resources created'
      requestBody:
        $ref: "#/components/requestBodies/SearchTransactionRequest"
      responses:
        '200':
          description: Transactions found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchTransactionResponse'
        '400':
          description: Formally invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemJson'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    ProblemJson:
      type: object
      properties:
        type:
          type: string
          format: uri
          description: |-
            An absolute URI that identifies the problem type. When dereferenced,
            it SHOULD provide human-readable documentation for the problem type
            (e.g., using HTML).
          default: about:blank
          example: https://example.com/problem/constraint-violation
        title:
          type: string
          description: |-
            A short, summary of the problem type. Written in english and readable
            for engineers (usually not suited for non technical stakeholders and
            not localized); example: Service Unavailable
        status:
          $ref: '#/components/schemas/HttpStatusCode'
        detail:
          type: string
          description: |-
            A human readable explanation specific to this occurrence of the
            problem.
          example: There was an error processing the request
        instance:
          type: string
          format: uri
          description: |-
            An absolute URI that identifies the specific occurrence of the problem.
            It may or may not yield further information if dereferenced.
    HttpStatusCode:
      type: integer
      format: int32
      description: |-
        The HTTP status code generated by the origin server for this occurrence
        of the problem.
      minimum: 100
      maximum: 600
      exclusiveMaximum: true
      example: 200
    SearchTransactionRequestFiscalCode:
      type: object
      description: Search transaction by user fiscal code
      properties:
        type:
          type: string
        userFiscalCode:
          type: string
          minLength: 16
          maxLength: 16
      required:
        - type
        - userFiscalCode
      example:
        type: "USER_FISCAL_CODE"
        userFiscalCode: "MRGHRN97L02C469W"
    SearchTransactionRequestEmail:
      type: object
      description: Search transaction by user fiscal code
      properties:
        type:
          type: string
        userEmail:
          type: string
          pattern: (?:[a-zA-Z0-9!#$%&'*+\/=?^_`{|}~-]+(?:\.[a-zA-Z0-9!#$%&'*+\/=?^_`{|}~-]+)*|"(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21\x23-\x5b\x5d-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])*")@(?:(?:[a-zA-Z0-9](?:[a-zA-Z0-9-]*[a-zA-Z0-9])?\.)+[a-zA-Z0-9](?:[a-zA-Z0-9-]*[a-zA-Z0-9])?|\[(?:(?:25[0-5]|2[0-4]\d|[01]?\d\d?)\.){3}(?:25[0-5]|2[0-4]\d|[01]?\d\d?|[a-zA-Z0-9-]*[a-zA-Z0-9]:(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21-\x5a\x53-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])+)\])
      required:
        - type
        - userEmail
      example:
        type: "USER_EMAIL"
        userEmail: "mario.rossi@pagopa.it"
    SearchTransactionResponse:
      type: object
      description: TransactionResponse
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResult'
        page:
          $ref: '#/components/schemas/PageInfo'
      required:
        - transactions
        - page
    SearchTransactionRequestRptId:
      type: object
      description: Search transaction by user fiscal code
      properties:
        type:
          type: string
        rptId:
          type: string
          pattern: ^([0-9]{29})$
      required:
        - type
        - rptId
      example:
        type: "RPT_ID"
        rptId: "77777777777302011111111111111"
    SearchTransactionRequestPaymentToken:
      type: object
      description: Search transaction by payment token
      properties:
        type:
          type: string
        paymentToken:
          type: string
      required:
        - type
        - paymentToken
      example:
        type: "PAYMENT_TOKEN"
        paymentToken: "paymentToken"
    SearchTransactionRequestTransactionId:
      type: object
      description: Search transaction by transaction id
      properties:
        type:
          type: string
        transactionId:
          type: string
          minLength: 32
          maxLength: 32
      required:
        - type
        - transactionId
      example:
        type: "TRANSACTION_ID"
        transactionId: "c9644451389e47b0a7d8e9d488fcd502"
    PmSearchTransactionRequest:
      type: object
      oneOf:
        - $ref: '#/components/schemas/SearchTransactionRequestFiscalCode'
        - $ref: '#/components/schemas/SearchTransactionRequestEmail'
      discriminator:
        propertyName: type
        mapping:
          USER_FISCAL_CODE: "#/components/schemas/SearchTransactionRequestFiscalCode"
          USER_EMAIL: "#/components/schemas/SearchTransactionRequestEmail"
    TransactionResult:
      type: object
      description: TransactionResponse
      properties:
        userInfo:
          $ref: '#/components/schemas/UserInfo'
        transactionInfo:
          $ref: '#/components/schemas/TransactionInfo'
        paymentInfo:
          $ref: '#/components/schemas/PaymentInfo'
        pspInfo:
          $ref: '#/components/schemas/PspInfo'
        product:
          $ref: '#/components/schemas/Product'
      required:
        - userInfo
        - transactionInfo
        - paymentInfo
        - pspInfo
        - product
    UserInfo:
      type: object
      description: User information
      properties:
        userFiscalCode:
          type: string
          minLength: 16
          maxLength: 16
        notificationEmail:
          type: string
          pattern: (?:[a-zA-Z0-9!#$%&'*+\/=?^_`{|}~-]+(?:\.[a-zA-Z0-9!#$%&'*+\/=?^_`{|}~-]+)*|"(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21\x23-\x5b\x5d-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])*")@(?:(?:[a-zA-Z0-9](?:[a-zA-Z0-9-]*[a-zA-Z0-9])?\.)+[a-zA-Z0-9](?:[a-zA-Z0-9-]*[a-zA-Z0-9])?|\[(?:(?:25[0-5]|2[0-4]\d|[01]?\d\d?)\.){3}(?:25[0-5]|2[0-4]\d|[01]?\d\d?|[a-zA-Z0-9-]*[a-zA-Z0-9]:(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21-\x5a\x53-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])+)\])
        surname:
          type: string
          maxLength: 512
        name:
          type: string
          maxLength: 512
        username:
          type: string
          maxLength: 128
        authenticationType:
          type: string
      example:
        userFiscalCode: "user_fiscal_code"
        notificationEmail: "test@test.it"
        surname: "Surname"
        name: "Name"
        username: "username"
        authenticationType: "auth type"
    TransactionInfo:
      type: object
      description: Transaction info
      properties:
        creationDate:
          type: string
          format: date-time
          description: transaction creation date
        status:
          type: string
        statusDetails:
          type: string
        events:
          type: array
          items:
            $ref: '#/components/schemas/EventInfo'
        eventStatus:
          type: string
          description: Possible statuses a transaction can be in
          enum:
            - ACTIVATED
            - AUTHORIZATION_REQUESTED
            - AUTHORIZATION_COMPLETED
            - CLOSURE_REQUESTED
            - CLOSED
            - CLOSURE_ERROR
            - NOTIFIED_OK
            - NOTIFIED_KO
            - NOTIFICATION_ERROR
            - NOTIFICATION_REQUESTED
            - EXPIRED
            - REFUNDED
            - CANCELED
            - EXPIRED_NOT_AUTHORIZED
            - UNAUTHORIZED
            - REFUND_ERROR
            - REFUND_REQUESTED
            - CANCELLATION_REQUESTED
            - CANCELLATION_EXPIRED
        amount:
          type: integer
          minimum: 0
          maximum: 999999999
          description: Amount for payments, in euro cents
        fee:
          type: integer
          minimum: 0
          maximum: 999999999
          description: Amount for payments, in euro cents
        grandTotal:
          type: integer
          minimum: 0
          maximum: 999999999
          description: Amount for payments, in euro cents
        rrn:
          type: string
        authorizationCode:
          type: string
        authorizationOperationId:
          type: string
          description: Operation id for authorization. Present only if the payment gateway is NPG
        refundOperationId:
          type: string
          description: Operation id for the refund. Present only if the payment gateway is NPG
        paymentMethodName:
          type: string
        brand:
          type: string
        authorizationRequestId:
          type: string
          description: Authorization request id
        paymentGateway:
          type: string
          description: Payment gateway used to perform transaction
        correlationId:
          type: string
          format: uuid
          description: correlation id for a transaction executed with NPG
        gatewayAuthorizationStatus:
          type: string
          description: payment gateway authorization status
        gatewayErrorCode:
          type: string
          description: payment gateway authorization error code
      example:
        creationDate: "2023-08-02T14:42:54.047"
        status: "status"
        statusDetails: "status detail"
        eventStatus: "NOTIFIED_OK"
        amount: 100
        fee: 10
        grandTotal: 110
        rrn: "rrn"
        authorizationCode: "auth code"
        authorizationOperationId: "009911013"
        refundOperationId: "645561643244"
        paymentMethodName: "payment method name"
        brand: "brand"
        authorizationRequestId: "authorizationRequestId"
        paymentGateway: "VPOS"
        correlationId: "30846e8f-efa1-47ad-abad-08cfb30e5c09"
        gatewayAuthorizationStatus: "DECLINED"
        gatewayErrorCode: "121"
    PaymentInfo:
      type: object
      description: Payment info
      properties:
        origin:
          type: string
        idTransaction:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/PaymentDetailInfo'
    PaymentDetailInfo:
      type: object
      description: Payment details
      properties:
        subject:
          type: string
        iuv:
          type: string
        rptId:
          type: string
        amount:
          $ref: '#/components/schemas/AmountEuroCents'
        paymentToken:
          type: string
        creditorInstitution:
          type: string
        paFiscalCode:
          type: string
      example:
        subject: "subject"
        iuv: "302001069073736640"
        rptId: "rptId"
        idTransaction: "paymentContextCode"
        paymentToken: "payment token"
        creditorInstitution: "66666666666"
        amount: 99999999
        paFiscalCode: "77777777777"
    PspInfo:
      type: object
      description: PSP info
      properties:
        pspId:
          type: string
        businessName:
          type: string
        idChannel:
          type: string
      example:
        pspId: "EXAMPLEPSP"
        businessName: "businessName"
        idChannel: "13212880150_02_ONUS"
    AmountEuroCents:
      description: Amount for payments, in euro cents
      type: integer
      minimum: 0
      maximum: 99999999
    PageInfo:
      description: Information about the returned query page
      type: object
      properties:
        current:
          type: integer
          description: Current returned page index (0-based)
        total:
          type: integer
          description: Total pages for the query (based on requested page size)
        results:
          type: integer
          description: Transactions returned into the current page
      required:
        - current
        - results
        - total
    Product:
      type: string
      enum: [ "PM", "ECOMMERCE" ]
      description: Product from which transaction belongs
    EcommerceSearchTransactionRequest:
      type: object
      oneOf:
        - $ref: '#/components/schemas/SearchTransactionRequestRptId'
        - $ref: '#/components/schemas/SearchTransactionRequestPaymentToken'
        - $ref: '#/components/schemas/SearchTransactionRequestTransactionId'
        - $ref: '#/components/schemas/SearchTransactionRequestEmail'
        - $ref: '#/components/schemas/SearchTransactionRequestFiscalCode'
      discriminator:
        propertyName: type
        mapping:
          RPT_ID: "#/components/schemas/SearchTransactionRequestRptId"
          PAYMENT_TOKEN: "#/components/schemas/SearchTransactionRequestPaymentToken"
          TRANSACTION_ID: "#/components/schemas/SearchTransactionRequestTransactionId"
          USER_EMAIL: "#/components/schemas/SearchTransactionRequestEmail"
          USER_FISCAL_CODE: "#/components/schemas/SearchTransactionRequestFiscalCode"
    EventInfo:
      type: object
      description: Event info
      properties:
        creationDate:
          type: string
          format: date-time
        eventCode:
          type: string
      example:
        creationDate: "2023-08-07T15:32:56.592837917Z[Etc/UTC]"
        eventCode: "TRANSACTION_ACTIVATED_EVENT"
    SearchMetricsRequest:
      type: object
      properties:
        clientId:
          type: string
          description: Filter by client ID
        pspId:
          type: string
          description: Filter by PSP ID
        paymentTypeCode:
          type: string
          description: Filter by payment type code
        timeRange:
          type: object
          required:
            - startDate
            - endDate
          properties:
            startDate:
              type: string
              format: date-time
              description: Start datetime (must be within one hour of endDate)
            endDate:
              type: string
              format: date-time
              description: End datetime (must be within one hour of startDate)
      required:
        - timeRange
        - clientId
        - pspId
        - paymentTypeCode
      example:
        clientId: "client-app-123"
        pspId: "psp-001"
        paymentTypeCode: "TYPE_ABC"
        timeRange:
          startDate: "2023-11-30T14:00:00.000Z"
          endDate: "2023-11-30T14:59:59.999Z"
    TransactionMetricsResponse:
      type: object
      description: Metrics grouped by transaction status (full map with all statuses)
      properties:
        ACTIVATED: { type: integer, minimum: 0 }
        AUTHORIZATION_REQUESTED: { type: integer, minimum: 0 }
        AUTHORIZATION_COMPLETED: { type: integer, minimum: 0 }
        CLOSURE_REQUESTED: { type: integer, minimum: 0 }
        CLOSED: { type: integer, minimum: 0 }
        CLOSURE_ERROR: { type: integer, minimum: 0 }
        NOTIFIED_OK: { type: integer, minimum: 0 }
        NOTIFIED_KO: { type: integer, minimum: 0 }
        NOTIFICATION_ERROR: { type: integer, minimum: 0 }
        NOTIFICATION_REQUESTED: { type: integer, minimum: 0 }
        EXPIRED: { type: integer, minimum: 0 }
        REFUNDED: { type: integer, minimum: 0 }
        CANCELED: { type: integer, minimum: 0 }
        EXPIRED_NOT_AUTHORIZED: { type: integer, minimum: 0 }
        UNAUTHORIZED: { type: integer, minimum: 0 }
        REFUND_ERROR: { type: integer, minimum: 0 }
        REFUND_REQUESTED: { type: integer, minimum: 0 }
        CANCELLATION_REQUESTED: { type: integer, minimum: 0 }
        CANCELLATION_EXPIRED: { type: integer, minimum: 0 }
      example:
        ACTIVATED: 12
        CLOSED: 45
        NOTIFIED_OK: 20
        EXPIRED: 0
        REFUNDED: 0
        CANCELED: 0
        EXPIRED_NOT_AUTHORIZED: 0
        UNAUTHORIZED: 0
        REFUND_ERROR: 0
        REFUND_REQUESTED: 0
        CANCELLATION_REQUESTED: 0
        CANCELLATION_EXPIRED: 0
        AUTHORIZATION_REQUESTED: 0
        AUTHORIZATION_COMPLETED: 0
        CLOSURE_REQUESTED: 0
        CLOSURE_ERROR: 0
        NOTIFIED_KO: 0
        NOTIFICATION_ERROR: 0
        NOTIFICATION_REQUESTED: 0
  requestBodies:
    SearchMetricsRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SearchMetricsRequest'
          examples:
            full example:
              value:
                clientId: "client-app-123"
                pspId: "psp-001"
                paymentTypeCode: "TYPE_ABC"
                timeRange:
                  startDate: "2023-11-30T14:00:00.000Z"
                  endDate: "2023-11-30T14:59:59.999Z"
            only clientId:
              value:
                clientId: "client-app-456"
                timeRange:
                  startDate: "2023-11-30T15:00:00.000Z"
                  endDate: "2023-11-30T15:45:00.000Z"
            only pspId:
              value:
                pspId: "psp-xyz"
                timeRange:
                  startDate: "2023-12-01T10:00:00.000Z"
                  endDate: "2023-12-01T10:50:00.000Z"
    EcommerceSearchTransactionRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EcommerceSearchTransactionRequest'
          examples:
            search by rpt id:
              value:
                type: RPT_ID
                rptId: "77777777777111111111111111111"
            search by payment token:
              value:
                type: PAYMENT_TOKEN
                paymentToken: "paymentToken"
            search by transaction id:
              value:
                type: TRANSACTION_ID
                transactionId: "transactionId"
            search by user email:
              value:
                type: USER_EMAIL
                userEmail: "test@test.it"
            search by user fiscal code:
              value:
                type: USER_FISCAL_CODE
                userFiscalCode: "user_fiscal_code"
    SearchTransactionRequest:
      required: true
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/PmSearchTransactionRequest'
              - $ref: '#/components/schemas/EcommerceSearchTransactionRequest'
            discriminator:
              propertyName: type
              mapping:
                USER_FISCAL_CODE: "#/components/schemas/SearchTransactionRequestFiscalCode"
                USER_EMAIL: "#/components/schemas/SearchTransactionRequestEmail"
                RPT_ID: "#/components/schemas/SearchTransactionRequestRptId"
                PAYMENT_TOKEN: "#/components/schemas/SearchTransactionRequestPaymentToken"
                TRANSACTION_ID: "#/components/schemas/SearchTransactionRequestTransactionId"
          examples:
            search by user fiscal code:
              value:
                type: USER_FISCAL_CODE
                userFiscalCode: "user_fiscal_code"
            search by user email:
              value:
                type: USER_EMAIL
                userEmail: "test@test.it"
            search by rpt id:
              value:
                type: RPT_ID
                rptId: "77777777777111111111111111111"
            search by payment token:
              value:
                type: PAYMENT_TOKEN
                paymentToken: "paymentToken"
            search by transaction id:
              value:
                type: TRANSACTION_ID
                transactionId: "transactionId"